FinishPdfService

+++ Fikra/Service/Interface/IPdfService.cs
@@ -2,7 +2,7 @@
 {
     public interface IPdfService
     {
-        public Task<string> GenerateContract(string ideaOwner, string investor, decimal budget, DateTime date,
+        public Task<string> GenerateContract(string ideaOwner, string investor, double budget, DateTime date,
                                   string ownerSignature, string investorSignature, byte[] logoBytes);
         public Task<byte[]> ReciveImage();
     }
